refactor: 🧑‍💻 move raws and directory management to context providers
